src/requirements.txt
paho-mqtt==1.6.1
adafruit-io==2.3.1
numpy==1.23.5
opencv-python==4.6.0.66
tensorflow==2.12.0
keras==2.12.0
Pillow==9.4.0
pyserial==3.5
imutils==0.5.4
CMake==3.26.3
dlib==19.24.1
face-recognition==1.3.0
firebase-admin==6.1.0
python-firebase==1.2